Update .gitignore with visualstudio, csharp, and dotnetcore templates
Expanded the .gitignore file by including templates for csharp, visualbasic, visualstudio, and dotnetcore. This will ensure that files and directories that are not needed in the repository, especially those generated by Visual Studio and those related to csharp and dotnetcore, are properly ignored, making the project cleaner and reducing unnecessary tracked files.