Add tail property and O(1) addNode method to GenericLinkedList
A tail property has been added to the GenericLinkedList class to enable constant time (O(1)) addition of nodes to the end of the list. This enhancement modifies the AddNode method and helps enhance the LinkedList's performance by reducing the time complexity from O(n) to O(1) for adding elements at the end of the list.