Implement multiple output methods for LinkedList
The PrintList method of the GenericLinkedList class has been adapted to allow for a variety of output targets, including Console, StreamWriter or StringWriter. The Node class has also been refactored and made private within the GenericLinkedList class, while the former version of the Node class has been marked as Obsolete. This change provides a cleaner interface and addresses output flexibility.