Rename LinkedList `Head` and `Tail` to `First` and `Last`
This commit renames `Head` and `Tail` properties of the `GenericLinkedList` class to `First` and `Last` respectively. Corresponding changes have been made in the test classes and the `Program.cs` in the LinkedList2 Directory to accommodate this refactoring. This terminology change improves code readability and aligns more closely with .NET's implementation of the LinkedList data structure.