Update GenericLinkedList and add ItemAt method
The List property's setter is set to private to maintain integrity. The "Last" reference is updated to the newly created node instead of pointing to the "First". Moreover, a new method ItemAt() is implemented to retrieve an item at a specified index. An error is thrown if the requested index is out of bounds or the list is empty.