Refactor generic list print function to extension method
The printing function for generic linked lists has been refactored from being part of the main function and is now located in the 'GenericLinkedListExtension.cs' file as an extension method. Additionally, all calls to this function within the main code have been commented out for now. This refactoring is part of a broader effort to improve code organization and readability.