Update variable declarations to use implicit typing
This commit updates variable declarations in several files, specifically in test files and in main programs, to use implicit typing instead of explicit. This was carried out to make the code more readable and concise, by leveraging .NET's type inference mechanisms such as the 'var' keyword in C# and removing explicit type declarations in VB where not necessary.